--- TechFlurry.Blazor.MVVM/Attributes/BindWithAttribute.cs
@@ -1,12 +1,0 @@
-﻿namespace TechFlurry.Blazor.MVVM.Attributes;
-
-[AttributeUsage(AttributeTargets.Property)]
-public class BindWithAttribute : Attribute
-{
-    public BindWithAttribute(string propertyName)
-    {
-        PropertyName = propertyName;
-    }
-
-    public string PropertyName { get; }
-}
