+++ TechFlurry.Blazor.MVVM/Setup.cs
@@ -1,8 +1,6 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.DependencyInjection.Extensions;
-using System.ComponentModel;
 using System.Reflection;
-using TechFlurry.Blazor.MVVM.Infrastructure;
+using TechFlurry.Blazor.MVVM.Utils.Extensions;
 using TechFlurry.Blazor.MVVM.ViewModels;
 
 namespace TechFlurry.Blazor.MVVM;
