+++ TechFlurry.Blazor.MVVM/ViewModels/ViewModelBase.cs
@@ -1,10 +1,9 @@
 ﻿using System.ComponentModel;
 using System.Runtime.CompilerServices;
-using TechFlurry.Blazor.MVVM.Infrastructure;
 
 namespace TechFlurry.Blazor.MVVM.ViewModels
 {
-    public interface IViewModelBase : INotifyPropertyChanged, IDisposable, IAsyncDisposable, IEquatable<IViewModelBase>
+    public interface IViewModelBase : INotifyPropertyChanged, IDisposable, IAsyncDisposable
     {
     }
 
@@ -16,33 +15,18 @@ namespace TechFlurry.Blazor.MVVM.ViewModels
         }
 
         public event PropertyChangedEventHandler? PropertyChanged;
-        public void Dispose()
+        public virtual void Dispose()
         {
+            GC.SuppressFinalize(this);
             //Model.Dispose ();
         }
         //method to raise the PropertyChanged event
-        internal void RaisePropertyChanged([CallerMemberName] string propertyName = "")
-        {
-            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-        }
-        //method to assign the value to the property and raise the PropertyChanged event
-        protected void SetMember<T>(ref T storage, T value, [CallerMemberName] string memberName = null)
-        {
-            if (!EqualityComparer<T>.Default.Equals(storage, value))
-            {
-                storage = value;
-                RaisePropertyChanged(memberName);
-            }
-        }
+        internal void RaisePropertyChanged([CallerMemberName] string propertyName = "") => PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
 
-        public ValueTask DisposeAsync()
+        public virtual ValueTask DisposeAsync()
         {
+            GC.SuppressFinalize(this);
             return ValueTask.CompletedTask;
-        }
-        public abstract bool Equals(IViewModelBase? other);
-        public override bool Equals(object? obj)
-        {
-            return obj is IViewModelBase ? Equals(obj as IViewModelBase) : base.Equals(obj);
         }
     }
 }
